now there are two ways to access the data on it: ->getGenus() and now ->getNotes()
 you can think about any relationship in two directions: each GenusNote has one Genus. Or, each Genus has many GenusNote. And in Doctrine, you can map just one side of a relationship, or both.
The notes property : This is the inverse side of the relationship. Above this, add a OneToMany annotation with targetEntity set to GenusNote and a mappedBy set to genus - that's the property in GenusNote that forms the main, side of the relation.
don't get confused: there's still only one relation in the database: but now there are two ways to access the data on it: ->getGenus() and now ->getNotes()

Add an inversedBy set to notes on this side: to point to the other property.