what happens when authentication is successful? It's awesome: the user is automatically redirected back to the last page they tried to visit before being forced to login
In other words, if the user tried to go to /checkout and was redirected to /login, then they'll automatically be sent back to /checkout so they can continue buying your awesome stuff.
But, in case they go directly to /login and there is no previous URL to send them to, we need a backup plan. That's the purpose of onAuthenticationSuccess()