Setting up the Entity User Provider
Delete the in_memory stuff and replace it with our_users: that's a totally meaningless machine name - it could be anything.
The property part is not something we care about right now, but we will use and talk about it later.
We can finally surf around the site and stay logged in.