Handling the Plain Password
we'll start with a plain text password, encrypt it through the bcrypt algorithm and store that on the password property.
The best way is to set the plain-text password on the User and encode it automatically via a Doctrine listener when it saves.
To do that, add a new property on User called plainPassword. But wait! Don't persist this with Doctrine: we will of course never store plain-text passwords.
This is just a temporary-storage place during a single request.