we'll use a Doctrine listener to read the plainPassword property, encode it, and update password. That means that password will be set to a value before it actually saves: it won't remain null.
Doctrine listeners are not called if Doctrine thinks that an object has not been updated.
If you eventually create a 'change password' form, then the only property that will be updated is plainPassword.
Since this is not persisted, Doctrine will think the object is un-changed, or clean.
In that case, the listeners will not be called, and the password will not be changed.

But by adding this line, the object will always look like it has been changed, and life will go on like normal.