we do need a password field
the property we want to set on User is not actually the password property.
We need to set plainPassword. Add this. It'll be a password type.
But, if you want the user to type the password twice, use a RepeatedType.
Then, in the third argument, pass the real type with type set to PasswordType::class.
That'll render two password boxes. And if the values don't match, validation will automatically fail.