The UniqueEntity Constraint
Check that email is already taken.
The options we need are fields - which tell it which field needs to be unique in the database - and message so we can say something awesome when it happens.
So add fields={email}. This is called fields because you could make this validation be unique across several columns.