we need this annotation to only work on the registration form
by default, all constraints live in a group called Default.
And when your form is validated, it validates all constraints in this Default group.
So now that we've put this into a different group called Registration, when the form validates, it doesn't validate using this constraint.

To use this annotation only on the registration form, we need to make that form validate everything in the Default group and the Registration group.