Automatically Login after Registration!
This is always pretty easy, but it's especially easy because we're using Guard authentication.
it redirected us back to the URL we were trying to visit before we were sent to the login page.
This is really great because it means that if your user tries to access a secure page - like your checkout form - they'll end up back on the checkout form after registration. Then they can keep buying your cool stuff.